feat: add retries when reading the port and check for empty file
For some reason, PIA might have multiple writes to the file. Some might delete it's content
We're adding checks if the file is empty, retry 5 times every 3 seconds.
If it's not empty, maybe it wrote some info but not all,
so we also retry reading the port 3 times every 3 seconds